Improvments for Router, Request and Response objects.
In the router, the first route matched handle the request.

Moved the send() code from a Response to the Router. This way, a body is
always completed when it matches a route.

Headers in a Request or Response getter to the Soup.MessageHeaders in the
respective message which provide much more flexibility.

Request objects have getter for their bodies and headers.