Working implementation for libsoup.
Request and Response are implemented using abstract class as they can
provide a part of the implementation.

This is necessary since FastCGI implementation will not use libsoup
Message and MessageHeaders. This provides is a good basis to build
FastCGI.

Using DataInputStream and DataOutputStream for body of Request and
Response as it is buffered and convenient for reading and writting
strings.