A matcher is a low-level utility to match Request.
Instead of relying on regular expression, it presents the matching
process as a simple callback applied on a request. If successful, the
request is filled with appropriate parameters.

This simplifies considerably the Route class, as a matcher is a closure
that can capture most if its work variable in the constructor.

matches and fire become simple wrapper around matcher and callback.