Implemented HTTP query.
Soup.URI does not provide a parsed HTTP query, so a VSGI.Request specify
a HashTable<string, string>? to represent the query.

Technically,

   /path/? should be an empty table
   /path/  should be a null table

This is the case for the Soup implementation, but the FastCGI one does
not provide the sufficent information to make that distinction.

Adapts the TestRequest to fit the new changes.

Adds an example demonstrating query possibilities.