Version the tests.
Versionning the tests will permit us to keep the framework
backward-compatible in minor release.

Technically, we shouldn't modify tests from a precedent minor release as
they have to pass if we haven't broken the API.

If we need to introduce a new behiaviour in the framework, we can create
a new test to do so and put the right version number.