Fixes non-null in recent FastCGI changes and cookie implementation.
The --enable-experimental-non-null cannot be stashed the way I thought because
waf does not do actual work in the function, but just defines dependencies.

I have commented it and the idea is to uncomment it when it's a good time to
cover null cases in the code.

Eventually, the examples will be strict too, so we could keep it enabled.