Generates versionned build files based on major and minor version.
Build files are versionned using MAJOR.MINOR pattern.

waf can use the name option to refer a target instead of the target option. This
is useful as we can now version the target name and reference the target using
its unversionned name.

The pkg-config flag will set the Cflags path based on major and minor version.

The -lfcgi compile flag is moved to the Libs.private for pkg-config.