Introduces middlewares for serving static content.
Support 'ETag' for serving files and resources and raise a
'304 Not Modified' status when appliable.

For 'GLib.File' delivery, the 'ETag' is obtained from the file
attributes.

For 'GLib.Resource' delivery, the 'ETag' is computed with SHA-1 and
stored in a cache to minimize the computational costs.

Replace the resource serving code in the example application and add a
link to a resource being served with the middleware.

Add minimal documentation.