Moves the scoping to the Route.
It is now the responsibility of the Route to scope its rule, rather than
processing it in the Router.

This will provide us a better flexibility, especially for the scoping regular
expression route.

Adapts the tests to suit that new behiaviour. Since the route handle the
scoping, we have to provide unscoped route as studs.