Push Map, Multimap and HashTable.
push_map, push_hashtable expects the type of the value to be GLib.Value?, so you
must box your data in order for it to work.

push_int_map, push_string_map, ... are providen if you have to push primitives.
This way, a Request query can be pushed in the environment.