Hold the application to serve.
Hold and release for each server request.

When a server start serving, it holds the application to keep it alive and
release it when it cannot handle anymore request.

This can be used to determine when the application is busy and do idle
processing.

FastCGI is triggered only when new data is available from its socket.