FastCGI support socket, port and default socket.
The previous implementation was buggy since it was not even considering the
default socket.

There is an implicit socket if it is not specified by FastCGI.open_socket. It is
generally used by server to spawn multiple process on multiple sockets.

Makes a distinction between socket and TCP port as it have different error
messages and eventually different options.

Backlog option for FastCGI if a socket or a TCP port is specified.