Real asynchronous requesting for FastCGI
FastCGI.request instance cannot be shared among threads, so we have to
initialize one per user request. It's more expensive, but it allow real
asychronous operations. Instead, a socket file descriptor is shared within the
process.

Request processing is triggered by an IOSource rather than busy waiting.

The FastCGIRequest implementation will finish and close its inner
FastCGI.request instance in its destructor, but not close the socket descriptor.
This way, the request will be closed even in asynchronous context.

The vapi was missing a close operation on FastCGI.Stream object and a
listen_sock attribute in FastCGI.request.

The FastCGIResponse was simplified and does not need a reference to the
FastCGI.request instance, just access to the out and err streams.