Accessing a mapped URL with a wrong method returns a 405
Instead of just returning a '404 File Not Found' when accessing a mapped URL
with a non-mapped method (e.g. GET /index/ is mapped, POST /index/ is accessed),
now a '405 Method Not Allowed' is thrown.

The 'Allow' header containing acceptable methods is included in the response.

This gives more valuable information and is more consistent with the HTTP
protocol specification.

Provide a simple test for the use case.