Fixes FastCGI implementation with asynchronous handler.
It is not possible to perform asynchronous operation on a response inside a
handler that is executed asynchronously already. The example was failing most of
the time due to that issue.

Improved the implementation since the Request does not need to wrap the
FastCGI.request anymore, but just the input FileStream.