Fixes build on Travis CI.
The handle_local_option signal is available since gio >=2.40, but travis runs
with 2.32. It's okay to use the command_line signal instead.

Preprocessor instructions to enable features available with more recent version
of the libraries.

 - handle CLI arguments if gio-2.0 (>=2.40)
 - add_main_options if gio-2.0 (>=2.42)
 - uses the new server API if libsoup-2.4 (>=2.38)

Travis will eventually migrate to Ubuntu 14.04, so it will be possible to build
for both environments.