General documentation improvments.
Explains what version of what dependencies provide additionnal features.

Improves the module section to be more readable and fix the code example.

Adds references to libgda, memcached, redis-glib, mongodb-glib and couchdb-glib
in the persistency document. Also removes the TODO code examples which are just
not working.

Improves the resource document for clarity and fix the wscript example.

Improves the router document with section about regular expression, matcher
callback and setup/teardown signals.

Improves the FastCGI document to be more concise and make lighttpd the solution
to develop an application.