Tests for Router, Route and View.
Tests for all Router built-in HTTP methods and ServerError thrown in a handler.

Test for creating a Route from a rule using an undefined type. Since it's
requiring a subprocess trap, it is conditionnaly built with glib (>=2.38).

Tests for View were added:

 - tests for view creation from string, path and stream
 - test for pushing a collection containing an unknown type
 - test for pushing string with push_value
 - test for push an unknown type with push_value

Removes non-working types from push_value, it seems like typeof does not work on
primitives or array of primitives.