Fixes a potential async bug with Soup implementation.
Once the server callback executed, the default behiaviour of the server is to
complete the message, but there might be some asynchronous processing in-going.

This fix just pause the I/O on the message so that while the server is taking
back the control, it is not transmitting the data to the client and let the
application do the processing.