Exposes more GObject properties for VSGI.
Uses a property for Request.params.

Exposes FastCGI Socket as well as the underlying Soup.Server and Soup.Message
for the Soup implementation. It provides much more features to the application
and the ability to take advantage of a particular implementation.

The Request reference in Response must be a regular one, since in some context,
only the reference to the Response is preserved and it would not be good to have
the Request freed at this time.