Provides write_head and write_head_async to write status line and headers.
Request and Response bodies are immuable and filtered directly by their
implementations, otherwise it cannot be guarantee that filtering streams are
closed before the connection closes.

Unifies status line and headers write operation under a single write_head and
provide an asynchronous version, write_head_async. Only build_head has to be
implemented, which return the data to be written in the stream.