steal_connection is available with libsoup-2.4 (>=2.50)
Fallbacks on a MemoryOutputStream on the response_body MessageBody of the
Message if steal_connection is not available. It already chunks the data, so
chunking is also conditional.

It is easier to maintain a default Response body implementation that does not
wrap the raw_body. Most planned implementations (CGI, FastCGI) already handle
chunking, so it will be much less code.