Introduces Next in Router.
Route.Handler has now a third argument which is a callback used to keep routing.

The setup signal does not serve any purpose, so it has been removed. It can be
implemented using a simple catch-all route that call next.

The teardown signal will eventually be removed in the 0.2.* release since we
will have the possibility to connect a callback to do the teardown processings.

Removes the timer examples as it requires the setup signal. The example will be
restored in the 0.2.* branch.

Replaces the setup and teardown documentation section by a section describing
the Next delegate and how it can be used to perform similar operations.

Tests for the next callback that cover the recursion of process_routing.