Status code handler.
The status handler is invoked with perform_routing, so next can be used to jump
to the next status handler for a given status. If nothing matches, the router
handler is used.

Status handler must set the status code in the response as it remained in the
exact same state it was when the exception were thrown.

It also captures generated NOT_FOUND and METHOD_NOT_ALLOWED to provide a uniform
approach at handling status codes.