Route with a null rule will catch all request paths.
The rule used will automatically default to "<any:path>", according to the
definition of the any by the router. The matched path is made available in the
request parameters under the "path" key.

Since the router is principally rule-oriented, this feature allow a much better
integration of middlewares and setup operations.

Adds a test to ensure it works properly.

Documents the functionnality.