Provides a CGI implementation.
The code from VSGI.FastCGI is moved into VSGI.CGI to build a common CGI
implementation that can be reused for CGI-like protocols. It will make
the implementation of SCGI much simpler.

The server implementation is very simple as it only requires to dispatch
a single request and quit. The connection simply hold and release the
server to ensure that behiaviour.

Adds an example and a sample lighttpd configuration file to run the CGI
binary.

Tests focus on CGI implementation to ensure correct environment variable
interpretation.

Adds a document to describe the CGI implementation and how to run the
lighttpd example.