Considers 0-sized chunk as ending chunk in ChunkedConverter.
When deferring a chunk write due to a lack of outbuf size, the next call to
convert would not set the INPUT_AT_END flag (but the call after would). The
inbuf size must be checked instead of that flag.

Initializes bytes_read to 0 and increment it each time a byte is transfeered
from inbuf to outbuf to avoid using a counter.