Removes timeout as it is not usable with the processing model.
There is no way to know when a request has been fully processed (or when
the connection is finalized), so releasing does not make any sense and
is currently not even done properly. The best thing to do is just not
provide that feature and rely on external process control.

For VSGI implementations:

 - libsoup-2.4 holds indefinately or just run synchronously with older
   server API
 - FastCGI implementation holds indefinately.

Removes the CLI option and references to the inactivity timeout in the
documentation.