Write the head when a status is handled to avoid an empty message.
The current implementation was producing no output when a status was
thrown as write_head was never invoked.

The behaviour is documented in 'docs/vsgi/index' to make things clearer.
It is a good behiaviour as it let the application control what's being
written to the client.

Adds tests for head_written when status are thrown.