Improvments for the documentation.
Updates the code example and feature set in the README.md.

Removes the section in docs/application about deploying with FastCGI or
on Heroku to keep the document simpler.

Adds a table in docs/installation for extra dependencies and describe
the --enable-threading flag.

Specifies that the 'master' branch is not a stable reference and that
developer should use the latest stable release instead.

Describes the error message passed as a status in
docs/redirection-and-error.

Adds a table for type to make things clearer in docs/route.

Adds a section about matching and handling middlewares in docs/router to
provide guidelines for future implementations.