Fixes 9 warnings for the VSGI implementations.
When an error occur during option parsing and initialization, just
return an status code instead of continuing.

Response makes a IOError check when it calls write_head on the first
body access.

The libsoup-2.4 Response implementation does base call for the body
property so that it does not have to deal with IOError when invoking
write_head.

Messages are printed in the ApplicationCommandLine with print and
printerr instead of using message and error as they are part of the CLI
session, not logs.