Enables threading unconditionally.
Removes the '--enable-threading' option and adds gthread-2.0 as a
mandatory dependency.

Removes '--enable-threading' from the RPM build process.

Updates the pkg-config file to include gthread-2.0 in the required
packages.

Removes the reference to --enable-threading in the documentation and
adds gthread-2.0 to the required dependencies.

Having optionnal threading is complicating the build process for no
perceivable benefit.