Fixes versionning around the framework.
There's a lot of version numbers around the framework and it makes
bumping a very tricky task.

Removes the server header from VSGI.Soup implementation, it's not
necessary and forces us to update a version number. It can still be
setted with '--server-header'.

Moves 'vsgi/application.vala' to 'vsgi/vsgi.vala' for more consistency
with the namespace.

Updates version for GIR annotations as it was missing from the last
minor version bump.

Uses subst to generate valum.spec with the right version number.

Adds a checklist in 'docs/hacking' about version bumping and what has to
be changed in the code.

Build files are generated without the API_VERSION considering that:

 - C header ends in valum-0.2/valum.h
 - library can be versionned with extensions like libvalum.so.0.1.0
 - GIR, pkg-config and other files are already versionned by following
   their naming conventions