Fixes the Docker build on Ubuntu 14.04 LTS.
Uses ubuntu:latest so that we can use the latest stable dependencies
during the build process.

Dependencies are installed correctly and the build process install the
framework in '/usr' like a production installation.

The example in 'examples/app' serves on all interface to make the app
available outside the container.

Fixes two conditional compilations for VSGI.Soup.

 - 'set_ssl_cert_file' is only available for libsoup-2.4 (>=2.48).
 - move 'ServerListenOptions' in SOUP_2_48 as the listen API is only
   available prior to libsoup-2.4 (>=2.48).

Removes usages of SimpleIOStream in favour of VSGI.Test.Connection to
stub connections.

 - no need to check for gio-2.0 (>=2.44)
 - less tricky build (recent VAPI provides the binding)
 - no need to instanciate tons of resizable memory streams