Uses a stack to store the state.
Replaces the state by a stack of state as it provides many advantages:

 - no need to propagate a state, just pushing it on the stack
 - no need for request parameters as they can be pushed

Request parameters are pushed in the stack as GLib.Parameter instead of
being added in the request parameters.

The 'MatcherCallback' may push states on the stack like request
parameters.

The 'HandlerCallback' receives the stack and may populate it for the
next handler.

The 'NextCallback' does not have to propagate a state anymore as the
stack is automatically propagated to the next handler.

Updates the tests to take the changes into consideration.