Use write_all instead of write for response head
The 'write_all' attempts to write all the provided buffer instead of
just trying as much as possible. This change guarantees that the head is
fully written or a failure condition is set.

Return the write status instead of the number of bytes written, which is
already known if the write is successful. The IOError should be
monitored for failure.

The 'head_written' state is bound directly to the result of 'write_all',
making the code a little cleaner.