Invert the routing process
Since the HTTP method have been moved to Route, there's no point to
maintain a mapping of methods to route queues in the Router.

No need to check if the mapping contains the method anymore, so the code
is much less complex.

Route method can be 'null' if matching the method does not apply. This
is used for status-based routes.

Use a simple array to check if a method is already allowed. The correct
data structure would be something like a 'GenericSet', but considering
that the array will contain a very small amount of items (~3), it's just
faster to make a linear lookup.

Modify the order of methods in 'Allow' header to respect the order of
declaration.