Standalone examples for external dependencies
This is part of #115 which aim to provide more code examples to cover as
many technologies as possible.

All examples using libsoup-2.4 listen to all interfaces with the '--all'
flag set on. It makes it possible to run examples within the Docker
container and more generally as live demo available to local users.

The app example has been updated to be more concise and

 - the code follows the order of declaration in the template
 - some basic features has been moved to router features

The Memcached code from the default app example has been removed in
favour of:

 - a standalone example with three endpoint: GET, PUT and DELETE
 - use of status codes to manage errors
 - use libmemcached.vapi directly from nemequ/vala-extra-vapis

The JSON example has been moved and build on its own.

The Lua example has been rewritten to use directly the LuaVM.