Add View.from_resources to load a template from global resources
The new function can load resources stored in the global scope, which
are accessed with 'resources_open_stream' and 'resources_peek_data'. It
is a handy shortcut for:

    var view = new View.from_stream (resources_open_stream ("...", ...));

It does not really require tests as it is just wrapping lower-level
calls.

Add a section about the new feature in the ctpl document.