Fix captures extraction in 'Route.from_regex'
Use a lazy quantifier when matching the regular expression in the
capture group so that it won't match over a following capture group.

Extract all captures with 'MatchInfo.next' to find all capture groups
instead of only extracting the first one.

Add a test to ensure that multiple captures work correctly.