VSGI loader for GModule-based applications
Provide a loader, namely 'vsgi-loader', for application written as a
GModule.  The syntax is similar to gunicorn and make it possible to pass
CLI arguments to the server.

    vsgi-loader module -- --port=3003

It loads the symbole from the module, and forward arguments and serve
the application. It does not work for CGI due to the one process per
request constraint.

The implementation arguments must be separated by a '--' to make a clear
distinction from 'vsgi' arguments.

The default server technology is libsoup-2.4 and another implementation
can be specified with the '--server' flag.

The directory where the shared library is stored can be specified with
the '--directory' flag, defaulting on system path.

It support a built-in reloader based on 'GLib.FileMonitor' and
automatically reload the shared library on change.

Update documentation for loadable applications

Improve the section by covering more aspects with code examples.