Provide the 'and', 'or', 'not' and 'noop' middlewares
It is used to build conjunction and disjunction of matching middlewares,
which is essential for more complex request interpretation.

Provide 'Route.and' to build conjunctive clauses with a builder pattern.

Provide the 'noop' middleware and use it as a default value for the
handling callback to allow the use of the following syntax:

    app.get ("")
        .and (or (accept ("text/html"), accept ("text/html+xml")))
        .then ((req, res) => { ... });

It is optimized away whenever possible.