Update to waf-1.14 that fixes #103
Build a static library in 'build/static' using the 'vala_dir' option to
avoid conflicts of generated files. It is used by the tests and
examples so that they can be executed without installing anything.

Travis CI does not need to install the shared library as tests run
against the static library.

The RPM specfile does not have to export 'LD_LIBRARY_PATH' anymore to
run the tests.