Fix potential bug with overlapping custom methods
The strategy prior to glib-2.0>=2.44 were to join all allowed methods
and perform a substring search with the requested method. The fix
consist of joining the allowed methods with a ',' character.

It works if we consider only the standard HTTP methods as none of their
concatenation overlaps to produce an existing method, but that's not
necessarly the case as we allow custom methods. It is fine to use a ','
since it is not a legal character.

The typical failure case would be to name a custom method 'PUTPOST', it
would not allow neither 'PUT' or 'POST', but it would contain both
substrings.