Fix warnings and perform better error handling
Perform better error handling for the examples by delegating them to the
'Router'.

Check for 'IOError' when writting or closing the body if an error
occured in 'invoke'. If an error occur at this point, it's likely
unrecoverable, but we set the status to 500.

The only remaining warnings are related to the FastCGI hack for
converting its errors codes into 'IOError' and two deprecated features
from libsoup-2.4.