Route.then preserve the matching stack
Ensure that when a route is defined with 'then', it preserves the same
stack from preceeding route that called 'next'.

The issue is that the same matching process is reapplied and might push
on the routing stack values that are already available.

It is not a problem with a classical 'next' since it is expected to go
through the matching process all over again.