SCGI: Warn and refuse if the request body is too big for the buffer
This is only expected to happen on system where

    sizeof (size_t) < sizeof (int64)

In this case, the request body cannot be held in the stream buffer and
will be seriously misinterpreted.

Move the condition-check for the number of available bytes in the stream
buffer within the 'content_length > 0' branch. It's not useful to check
it otherwise.

Also warn and refuse if the stream fails to buffer the body.