Improve initialization in 'Router.handle'
Do not assign the response status to '200 OK', it's already done by
VSGI.

Since 'VSGI.Soup.Response' overwrites the 'status' property, it has to
be set manually in the server handler.

Do not initialize the 'Content-Type' header: under many circumstances
(like #143), it is better to have minimal headers.

The 'Content-Type' should be set explicitly, it's wrong to assume it
defaults to HTML.

Update tests and example to honor the new behaviour.