Improve content negociation middlewares
Fuzzy matching is provided by submitting a custom compare function to
'negociate'.

In short,

 - 'accept' handle '*/*' and 'type/*' patterns.
 - 'accept_charset', 'accept_encoding' and 'accept_language' handle '*'

Set the corresponding response header when a negociation succeeds.

If a charset is supplied to 'accept', it is negociated with
'accept_charset'.

Provide 'NegociateFlags.FINAL' to raise a 'ClientError.NOT_ACCEPTABLE'
if the negociation fails.

Provide minimal tests for 'negociate' and 'accept'.