Use 'GLib.Environ' utilities for CGI environment variables
The variables are represented as a simple 'KEY=VALUE' array, avoiding
memory overhead from the 'GLib.HashTable' and potentially better access
times.

CGI and FastCGI implementations already encode their environment this
way, so the code is much cleaner.

Update the tests to the new changes. The equality between the
environment cannot be asserted since it's being copied over.