Improve status handling
Feedback the error of a failed status handler with 'Router.invoke'
instead of applying a double try-catch.

Replace headers that are not multiple like 'Location' or
'WWW-Authenticate'.

Explicitly set the content encoding to 'Soup.Encoding.NONE' if the
status must not include an entity.

Cover all status codes described in HTTP/1.1 specification.