Provide a routing context instead of a stack
The context data structure consist of a tree of mapping from 'string' to
'Value?'. It will prepare the ground for typed parameters.

The number of stacked context corresponds to the depth of the routing.

Key existence and value is resolved recursively: each layer expose the
sum of itself and its parents.

Remove 'Request.params', it's not needed anymore as they can be obtained
directly from the context.

Update tests, documentation and examples.