Raise a '406 Not Acceptable' by default in 'negotiate'
As we can handle quality values, it's not useful to call 'next' by
default. Instead, a '406 Not Acceptable' is raised if none of the
expectations can be satisfied.

Replace 'NegotiateFlags.FINAL' by 'NegotiateFlags.NEXT' to call 'next'
instead of raising a '406 Not Acceptable' status.

Provide a 'forward' function to serve as a default value. It let
end-user write the following:

    app.use (accept ({"application/json"}));

Update tests to take these changes into consideration.