vsgi: Add a check for unwritten head
Assert that the response head has been written if its reference is not
held anywhere else by checking the refcount.

It is an error not to call 'write_head' and there's no possible mean to
actually make that call implicitly.

Yet, this does not guarantee that the head will eventually be written,
but it will help for most cases.