vsgi: Make 'ApplicationCallback' return true if handled
The 'ApplicationCallback' delegate now indicate if the request has been
handled by returning a truth value.

The possible scenarios where it is permitted to return 'true' are:

 - 'write_head' has been called at least once
 - 'write_head' will eventually be called (asynchronously)

Depending on the implementation, the sever can decide what to do next
such as raising a '501 Not Implemented', trying to satisfy the request
elsewhere or notify upstream that the request has not been handled.

Update tests and examples by taking care to use 'write_all' or 'next'
return value when possible.