vsgi: Add a 'content_length' option to 'convert'
It makes it possible to specify the resulting value (or removal) of the
'Content-Length' header when the converter is applied.

If the payload is already chunked, specifying an undetermined length
does not affect the transport encoding.

It fixes most issues where the payload size would become incorrect once
a converter is applied as it defaults to setting 'Encoding.EOF'.

Remove redundant 'set_encoding' in 'decode' middleware.

Update the documentation and add some tests convering particular cases.