vsgi: Provide a 'Connection' class inheriting from 'GLib.IOStream'
Having a 'Connection' provide access to the 'Server' which initiated a
'Request', which can be very handy. More generally, this can be used for
performing uninterruptible operations with 'hold' and 'release':

    app.get ("/", (req, res) => {
        req.connection.server.hold ();
        // perform uninterruptible operation
        req.connection.server.release ();
    });

There are also various advantages:

 - access to the DBus service
 - access the resource basepath
 - monitor the server itself

Implementations now expose a 'Connection' type to the outside world.

Remove 'using Soup;' to avoid name clashes and introduce 'Soup'
namespace where required.