vsgi: Rewrite with 'GLib.SocketAddress'
This is a revamp of the listen APIs that provide a much better
integration with GIO.

 - change 'listen' to take a 'GLib.SocketAddress'
 - if the 'listen' argument is 'null', fallback on the default interface
   for the protocol
 - add 'listen_socket' to listen on an already open socket
 - mark unsupported features with 'GLib.IOError.NOT_SUPPORTED'

It makes the 'Server' class much easier to play with programmatically.

To pass runtime options, just use GObject-style constructions:

    var https_server = Server.new ("http", https: true);

Move options to 'VSGI.Application' This way, it's the only class that
bother with CLI-parsing.

Make '--port', '--socket' and '--file-descriptor' arrays to listen on
multiple interfaces simultaneously.

Rename '--all' for '--any' to be more consistent with 'GLib.InetAddress.any'.

Let 'Server.uris' be owned by the callee so we don't have to worry for
it being possibly mutated externally.

Update tests and examples accordingly.