vsgi: Move 'GLib.Application' outside 'Server'
Instead, provide 'VSGI.Application' to manage a running 'Server'
instance. This way, the server can be run with a simple main loop or
more conveniently with log handling and CLI parsing.

The 'Server' constructor cannot take the application identifier anymore,
so it had to be removed.

Add 'Server.get_listen_options' to let the application add any relevant
options. It removes a lot of check for 'gio-2.0 >= 2.40' that we can
move to 'VSGI.Application'.

Update the CGI implementation which cannot rely anymore on hold/release
to quit. Instead, we only check if any sources are pending or exit.

Update examples and documentation.